Restructure application and add API routes
Renamed the 'v1' directory to 'api' and added API routes for reservations and tables using FastAPI routers. Also, psycogp2-binary is added to the requirements.txt, which is a PostgreSQL database adapter for Python. This restructuring will make for more straightforward API versioning in the future and implementation of routing system for handling different endpoints.